refactor(team): extract bridge handler, pane capture, cursors, provider binding
Staff-reviewer flagged that broker.go was still doing too much for a
"core" file. Five extractions:

(1) broker_bridge.go (141 LOC): handleBridge — the CEO-only
    cross-channel bridging endpoint. Domain handler mechanically
    identical to handlers in broker_office_channels.go /
    broker_messages.go; sat in broker.go only because nobody had
    picked a home for it. Wire shape: POST /bridge with actor +
    source + target + summary -> records signal + decision +
    automation message + action.

(2) broker_pane.go (96 LOC): capturePaneActivity — tmux pane
    activity capture orchestration. Iterates manifest members or
    a single 1:1 override, runs `tmux capture-pane -p -J`, diffs
    against b.lastPaneSnapshot, returns a slug -> last-5-meaningful-
    lines map for any agent whose pane content changed. Was
    orthogonal to broker lifecycle; now lives in its own file.

(3) broker_cursors.go (43 LOC): NotificationCursor /
    SetNotificationCursor / InsightsCursor / SetInsightsCursor.
    Two high-water-mark timestamps used by the broker for
    incremental sync (don't re-fan-out old messages, don't
    re-extract old insights). Tiny but cohesive — the four
    methods are the "incremental sync persistence cursor" theme.

(4) broker_provider_binding.go (62 LOC): SetMemberProvider /
    MemberProviderBinding / MemberProviderKind. The launcher
    dispatch switch consults these to decide which runtime
    (Claude / Codex / OpenClaw) to invoke per agent. Extracted so
    the binding contract is reviewable in isolation.

  - internal/team/broker.go: -255 LOC (1462 -> 1207)
  - 4 new files: 342 LOC

Cumulative broker.go reduction:
  3404 -> 1207
  Net: -2197 LOC; broker.go shrunk by 65%.

What stays in broker.go: types block (channelMessage / humanInterview
/ teamTask / officeMember / ... ~320 LOC), constructor (NewBroker /
NewBrokerAt), Start / StartOnPort / Stop / Reset, session-mode
setters, openclaw bridge attach, findChannelLocked /
ensureDMConversationLocked / findMemberLocked /
rebuildMemberIndexLocked, appendActionLocked. The remaining file is
the genuine core: types + lifecycle + locked-access primitives.

Tests pass.
