+++ internal/team/broker_pane.go
@@ -1,0 +1,96 @@
+package team
+
+import (
+	"context"
+	"fmt"
+	"os/exec"
+	"strings"
+
+	"github.com/nex-crm/wuphf/internal/company"
+)
+
+// Tmux pane activity capture. The activity watchdog calls this on a
+// timer to detect "is this agent doing something or stuck?" by
+// diffing the captured pane content against the last snapshot.
+//
+// In office mode it walks every member of the manifest and checks
+// the pane at the conventional position (wuphf-team:team.<i>); in
+// 1:1 mode it captures only pane 1. Returns a slug→summary map of
+// agents whose pane content changed since the last call (idle
+// agents are silently dropped).
+
+// capturePaneActivity returns the tmux pane content per agent slug.
+// If content is the same as last time, agent is idle — return nothing.
+func (b *Broker) capturePaneActivity(slugOverride string) map[string]string {
+	result := make(map[string]string)
+
+	type paneCheck struct {
+		slug   string
+		target string
+	}
+
+	var checks []paneCheck
+	if slugOverride != "" {
+		// 1:1 mode: only check pane 1
+		checks = append(checks, paneCheck{slug: slugOverride, target: fmt.Sprintf("%s:team.1", SessionName)})
+	} else {
+		manifest := company.DefaultManifest()
+		loaded, loadErr := company.LoadManifest()
+		if loadErr == nil && len(loaded.Members) > 0 {
+			manifest = loaded
+		}
+		for i, agent := range manifest.Members {
+			checks = append(checks, paneCheck{
+				slug:   agent.Slug,
+				target: fmt.Sprintf("wuphf-team:team.%d", i+1),
+			})
+		}
+	}
+
+	b.mu.Lock()
+	if b.lastPaneSnapshot == nil {
+		b.lastPaneSnapshot = make(map[string]string)
+	}
+	b.mu.Unlock()
+
+	for _, check := range checks {
+		paneOut, err := exec.CommandContext(context.Background(), "tmux", "-L", "wuphf", "capture-pane",
+			"-p", "-J",
+			"-t", check.target).CombinedOutput()
+		if err != nil {
+			continue
+		}
+
+		content := string(paneOut)
+
+		// Compare with previous snapshot
+		b.mu.Lock()
+		prev := b.lastPaneSnapshot[check.slug]
+		b.lastPaneSnapshot[check.slug] = content
+		b.mu.Unlock()
+
+		if content == prev {
+			// No change — agent is idle
+			continue
+		}
+
+		// Content changed — agent is active. Extract last 5 meaningful lines.
+		lines := strings.Split(content, "\n")
+		var meaningful []string
+		for i := len(lines) - 1; i >= 0 && len(meaningful) < 5; i-- {
+			trimmed := strings.TrimSpace(lines[i])
+			if trimmed == "" {
+				continue
+			}
+			meaningful = append(meaningful, trimmed)
+		}
+		// Reverse to chronological order
+		for i, j := 0, len(meaningful)-1; i < j; i, j = i+1, j-1 {
+			meaningful[i], meaningful[j] = meaningful[j], meaningful[i]
+		}
+		if len(meaningful) > 0 {
+			result[check.slug] = strings.Join(meaningful, "\n")
+		}
+	}
+	return result
+}
